A good place to save
Random Drink Screen has all the basic functionalities that I was intending to code. Now just need to pretty it up, and maybe add a few other features I thought could be fun to do. This gives a good showcase to friends and relatives of the idea of the app,